requirements.txt
torch
sentence-transformers
python-dotenv
pinecone 